% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneousInStock.R
\name{miscellaneousInStockUI_bottom}
\alias{miscellaneousInStockUI_bottom}
\title{生成区域}
\usage{
miscellaneousInStockUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
miscellaneousInStockUI_bottom()
}
