% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneousInStock.R
\name{miscellaneousInStockUI_left}
\alias{miscellaneousInStockUI_left}
\title{请输入文件}
\usage{
miscellaneousInStockUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
miscellaneousInStockUI_left()
}
