% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneousOutStock.R
\name{miscellaneousOutStockUI_right}
\alias{miscellaneousOutStockUI_right}
\title{生成区域}
\usage{
miscellaneousOutStockUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
miscellaneousOutStockUI_bottom()
}
