% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReceiveNotice.R
\name{purchaseReceiveNoticeUI_bottom}
\alias{purchaseReceiveNoticeUI_bottom}
\title{生成区域}
\usage{
purchaseReceiveNoticeUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
purchaseReceiveNoticeUI_bottom()
}
