% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReceiveNotice.R
\name{purchaseReceiveNoticeUI_left}
\alias{purchaseReceiveNoticeUI_left}
\title{请输入文件}
\usage{
purchaseReceiveNoticeUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
purchaseReceiveNoticeUI_left()
}
