% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReturnNotice.R
\name{purchaseReturnNoticeUI_bottom}
\alias{purchaseReturnNoticeUI_bottom}
\title{生成区域}
\usage{
purchaseReturnNoticeUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
purchaseReturnNoticeUI_bottom()
}
