% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReturnNotice.R
\name{purchaseReturnNoticeUI_left}
\alias{purchaseReturnNoticeUI_left}
\title{请输入文件}
\usage{
purchaseReturnNoticeUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
purchaseReturnNoticeUI_left()
}
