% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReturnStock.R
\name{purchaseReturnStockUI_left}
\alias{purchaseReturnStockUI_left}
\title{请输入文件}
\usage{
purchaseReturnStockUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
purchaseReturnStockUI_left()
}
