% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purchaseReturnStock.R
\name{purchaseReturnStockUI_right}
\alias{purchaseReturnStockUI_right}
\title{生成区域}
\usage{
purchaseReturnStockUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
purchaseReturnStockUI_bottom()
}
