% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleDeliveryNotice.R
\name{saleDeliveryNoticeUI_bottom}
\alias{saleDeliveryNoticeUI_bottom}
\title{生成区域}
\usage{
saleDeliveryNoticeUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
saleDeliveryNoticeUI_bottom()
}
